requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[aiosqlite]
pydantic-settings  
python-dotenv
pytest 
python-jose
python-multipart
passlib[bcrypt]
